+++ lib/lifecycle_logger.dart
@@ -1,17 +1,28 @@
 import 'package:flutter/widgets.dart';
 
 import 'src/app_lifecycle_observer.dart';
+import 'src/logger.dart';
+import 'src/route_lifecycle_observer.dart';
 
 export 'src/widget_lifecycle_mixin.dart' show LifecycleAware;
+export 'src/lifecycle_event.dart' show LifecycleEvent, LifecycleEventType;
+export 'src/logger.dart' show LifecycleEventSink;
+export 'src/route_lifecycle_observer.dart' show LifecycleRouteObserver;
 
 class LifecycleLogger {
   LifecycleLogger._();
 
   static AppLifecycleObserver? _observer;
+  static final LifecycleRouteObserver _routeObserver = LifecycleRouteObserver();
   static bool _attached = false;
+
+  static NavigatorObserver get routeObserver => _routeObserver;
 
   static void attach({
     bool debugOnly = true,
+    bool enableRouteObserver = false,
+    bool logToConsole = true,
+    LifecycleEventSink? sink,
     void Function()? onResume,
     void Function()? onPause,
     void Function()? onInactive,
@@ -20,6 +31,9 @@ class LifecycleLogger {
     if (debugOnly && !_isDebugMode) {
       return;
     }
+
+    LifecycleLog.configure(sink: sink, logToConsole: logToConsole);
+    _routeObserver.enabled = enableRouteObserver;
 
     if (_attached) {
       return;
@@ -44,6 +58,8 @@ class LifecycleLogger {
     WidgetsBinding.instance.removeObserver(_observer!);
     _observer = null;
     _attached = false;
+    _routeObserver.enabled = false;
+    LifecycleLog.reset();
   }
 
   static bool get _isDebugMode {